créa page inscription restaurant + client

+++ src/main/java/com/projet/heisenburger/repository/PlatRepository.java
@@ -5,7 +5,7 @@ import com.projet.heisenburger.model.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PlatRepository extends JpaRepository<Plat, Long> {
+public interface PlatRepository extends JpaRepository<Plat, Integer> {
 
     // Méthode pour trouver tous les plats appartenant à un restaurant spécifique
     List<Plat> findByRestaurant(Restaurant restaurant);
